chore: move constants to utils
