+++ src/lerobot/utils/random_utils.py
@@ -23,8 +23,8 @@ import numpy as np
 import torch
 from safetensors.torch import load_file, save_file
 
-from lerobot.constants import RNG_STATE
 from lerobot.datasets.utils import flatten_dict, unflatten_dict
+from lerobot.utils.constants import RNG_STATE
 
 
 def serialize_python_rng_state() -> dict[str, torch.Tensor]:
