+++ tests/optim/test_optimizers.py
@@ -14,10 +14,6 @@
 import pytest
 import torch
 
-from lerobot.constants import (
-    OPTIMIZER_PARAM_GROUPS,
-    OPTIMIZER_STATE,
-)
 from lerobot.optim.optimizers import (
     AdamConfig,
     AdamWConfig,
@@ -25,6 +21,10 @@ from lerobot.optim.optimizers import (
     SGDConfig,
     load_optimizer_state,
     save_optimizer_state,
+)
+from lerobot.utils.constants import (
+    OPTIMIZER_PARAM_GROUPS,
+    OPTIMIZER_STATE,
 )
